Fixes for Il8n and new default_error_messages
shoulda wasn't passing in the attribute to the new pretty versions of Il8n
error messages, so it was testing, against the literal strings:
"Please enter a value for {{attribute}}"

instead of:
"Please enter a value for Myfield"

and obviously it was breaking.

This now works - but there's no shoulda tests for it yet.